requirements.txt
# LifeBot 의존성

# 테스팅
pytest>=7.4.0
pytest-cov>=4.1.0

# 설정 관리
pyyaml>=6.0
python-dotenv>=1.0.0

# CLI
rich>=13.0.0
click>=8.1.0

# LLM (Phase 3 - 선택적)
openai>=1.0.0
anthropic>=0.8.0

# 웹 UI (Phase 4)
streamlit>=1.28.0
plotly>=5.17.0
pandas>=2.0.0
